requirements.txt
streamlit==1.52.2
google-genai==1.56.0
pillow
python-dotenv
requests
google-api-core
openai
fal-client
